+ 4.7.4 + 下载页面支持从镜像地址优先下载 * 更新一些翻译

+++ app/src/main/java/com/surcumference/fingerprint/network/updateCheck/github/bean/GithubAssetsInfo.java
@@ -13,5 +13,5 @@ public class GithubAssetsInfo {
     @SerializedName("browser_download_url")
     public String url;
 
-    public int size;
+    public long size;
 }
